+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/RequestParameters.kt
@@ -9,7 +9,7 @@ interface RequestParameters {
     val issuer: String?
     val audience: String?
     val state: String?
-    val transactionData: Set<String>?
+    val transactionData: Set<TransactionData>?
 
     /**
      * Reads the [OpenIdConstants.ClientIdScheme] of this request either directly from [clientIdScheme],