+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/Json.kt
@@ -7,10 +7,16 @@ import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.AuthorizationDetails
 import at.asitplus.openid.OpenIdAuthorizationDetails
 import at.asitplus.openid.RequestParameters
+import at.asitplus.openid.TransactionData
+import at.asitplus.rqes.collection_entries.QCertCreationAcceptance
+import at.asitplus.rqes.collection_entries.QesAuthorization
+import at.asitplus.rqes.serializers.Base64URLTransactionDataSerializer
 import at.asitplus.rqes.serializers.InputDescriptorSerializer
 import at.asitplus.rqes.serializers.RequestParametersSerializer
+import kotlinx.serialization.builtins.SetSerializer
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.modules.SerializersModule
+import kotlinx.serialization.modules.contextual
 import kotlinx.serialization.modules.polymorphic
 
 private val inputDescriptorModule = SerializersModule {
@@ -56,6 +62,13 @@ private val authorizationDetailsModule = SerializersModule {
     }
 }
 
+private val transactionDataModule = SerializersModule {
+    polymorphic(TransactionData::class) {
+        subclass(QesAuthorization::class, QesAuthorization.serializer())
+        subclass(QCertCreationAcceptance::class, QCertCreationAcceptance.serializer())
+    }
+}
+
 /**
  * This module fully overrides the one specified in `at.asitplus.openid.Json.kt`
  * It allows de-/serialization of open interfaces defined in this module as well as the OpenId module
@@ -64,6 +77,8 @@ private val extendedOpenIdSerializerModule = SerializersModule {
     include(inputDescriptorModule)
     include(requestParametersModule)
     include(authorizationDetailsModule)
+    include(transactionDataModule)
+    contextual(Base64URLTransactionDataSerializer)
 }