+++ vck-rqes/src/commonMain/kotlin/at/asitplus/wallet/lib/rqes/helper/OpenIdRqesParameters.kt
@@ -1,10 +1,12 @@
 package at.asitplus.wallet.lib.rqes.helper
 
-import at.asitplus.rqes.collection_entries.TransactionData
+import at.asitplus.openid.TransactionData
 
 /**
  * Wrapper to distinguish RQES related [AuthenticationRequestParameter] members better
  */
+
+@Deprecated("Replaced", ReplaceWith("OpenIdRequestOptions"))
 data class OpenIdRqesParameters(
     val transactionData: Set<TransactionData>,
 )